修复番剧状态数据int越界

+++ src/BiliLite.UWP/Models/Common/Season/SeasonDetailStatModel.cs
@@ -2,20 +2,20 @@
 {
     public class SeasonDetailStatModel
     {
-        public int Coins { get; set; }
+        public long Coins { get; set; }
 
-        public int Danmakus { get; set; }
+        public long Danmakus { get; set; }
 
-        public int Favorites { get; set; }
+        public long Favorites { get; set; }
 
         public string Followers { get; set; }
 
         public string Play { get; set; }
 
-        public int Reply { get; set; }
+        public long Reply { get; set; }
 
-        public int Share { get; set; }
+        public long Share { get; set; }
 
-        public int Views { get; set; }
+        public long Views { get; set; }
     }
 }
